fix: Workflow path is enumerated, but does not exist any longer
When workflow definition files are deleted, but had runs in the past,
the GitHub API still enumerates them as if they would exist. In reality,
however, they are gone.

This detail, however, only becomes apparent when accessing the workflow,
so this patch compensates for corresponding 404 errors returned by the
API.